Complete Bonus

+++ Assignment6/Assignment6/Plane.cpp
@@ -12,7 +12,7 @@ void createPlane()
 		-side, 0.0, -side, 1.0,
 	};
 
-	GLshort plane_normals[] = {
+	GLfloat plane_normals[] = {
 		0.0f, 1.0f, 0.0f,
 		0.0f, 1.0f, 0.0f,
 		0.0f, 1.0f, 0.0f,
@@ -23,22 +23,6 @@ void createPlane()
 		0, 1, 2,
 		2, 3, 0
 	};
-
-	//glGenVertexArrays(1, &plane_vao);
-	//glBindVertexArray(plane_vao);
-
-	//unsigned int handle[2];
-	//glGenBuffers(2, handle);
-
-	//glBindBuffer(GL_ARRAY_BUFFER, handle[0]);
-	//glBufferData(GL_ARRAY_BUFFER, sizeof(plane_vertices), plane_vertices, GL_STATIC_DRAW);
-	//glVertexAttribPointer((GLuint)0, 4, GL_FLOAT, GL_FALSE, 0, 0);
-	//glEnableVertexAttribArray(0);  // Vertex position
-
-	//glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, handle[1]);
-	//glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(plane_indices), plane_indices, GL_STATIC_DRAW);
-
-	//glBindVertexArray(0);
 
 
 	glGenVertexArrays(1, &plane_vao);
